Move output worker to ouput (#4)
